Prototype of Bill Calculation of Hospital
This console application consist of 5 classes. These classes perform operations inside to calculate pharmacy charges, room charges and doctor's fees. Doctor fees is differs according to doctor is psychiatric or not It has test function inside the project. It is an example of calculating charges of hospital with displaying patients informations also.